+++ DataRepo/tests/models/test_tracer.py
@@ -42,6 +42,7 @@ class TracerTests(TracebaseTestCase):
         each TracerLabel record creation, after which it has its final value.
         """
         # Throws DoesNotExist exception if not found
-        self.assertTrue(MaintainedModel.are_autoupdates_enabled())
         to = create_tracer_record()
+        coordinator = MaintainedModel._get_current_coordinator()
+        self.assertTrue(coordinator.are_autoupdates_enabled())
         self.assertEqual("glucose-[2,3-13C2,4-17O1]", to.name)