+++ DataRepo/tests/views/test_protocol_views.py
@@ -9,6 +9,22 @@ from DataRepo.models.maintained_model import (
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
 
 
+def assert_coordinator_state_is_initialized():
+    # Obtain all coordinators that exist
+    all_coordinators = [MaintainedModel._get_default_coordinator()]
+    all_coordinators.extend(MaintainedModel._get_coordinator_stack())
+    if 1 != len(all_coordinators):
+        raise ValueError(
+            f"Before setting up test data, there are {len(all_coordinators)} MaintainedModelCoordinators."
+        )
+    if all_coordinators[0].auto_update_mode != "immediate":
+        raise ValueError(
+            "Before setting up test data, the default coordinator is not in immediate autoupdate mode."
+        )
+    if 0 != all_coordinators[0].buffer_size():
+        raise UncleanBufferError()
+
+
 class ProtocolViewTests(TracebaseTestCase):
     """
     Test two list views for subsets of protocols
@@ -20,8 +36,7 @@ class ProtocolViewTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
         # Ensure the auto-update buffer is empty.  If it's not, then a previously run test didn't clean up after itself
-        if MaintainedModel.buffer_size() > 0:
-            raise UncleanBufferError()
+        assert_coordinator_state_is_initialized()
 
         call_command("load_study", "DataRepo/example_data/test_dataframes/loading.yaml")