+++ DataRepo/utils/sample_table_loader.py
@@ -11,6 +11,7 @@ from DataRepo.models import (
     AnimalLabel,
     FCirc,
     Infusate,
+    MaintainedModel,
     Protocol,
     Sample,
     Study,
@@ -20,7 +21,6 @@ from DataRepo.models.hier_cached_model import (
     disable_caching_updates,
     enable_caching_updates,
 )
-from DataRepo.models.maintained_model import MaintainedModel
 from DataRepo.models.researcher import (
     UnknownResearcherError,
     get_researchers,
@@ -142,11 +142,11 @@ class SampleTableLoader:
     def __init__(
         self,
         sample_table_headers=DefaultSampleTableHeaders,
-        validate=False,  # DO NOT USE MANUALLY - THIS WILL NOT ROLL BACK UPON ERROR (handle in outer atomic transact)
+        validate=False,  # Only affects what is/isn't a warning
         verbosity=1,
         skip_researcher_check=False,
         defer_autoupdates=False,
-        defer_rollback=False,
+        defer_rollback=False,  # DO NOT USE MANUALLY - THIS WILL NOT ROLL BACK (handle in atomic transact in caller)
         dry_run=False,
     ):
         # Header config
@@ -157,11 +157,9 @@ class SampleTableLoader:
         self.verbosity = verbosity
         self.dry_run = dry_run
         self.validate = validate
-
         # How to handle mass autoupdates
         self.defer_autoupdates = defer_autoupdates
-
-        # Whether to rollback upon error or defer it to the caller
+        # Whether to rollback upon error or keep the changes and defer rollback to the caller
         self.defer_rollback = defer_rollback
 
         # Caching overhead
@@ -200,52 +198,40 @@ class SampleTableLoader:
             "TIME_COLLECTED": ["m", "min", "mins", "minute", "minutes"],
         }
 
+    @MaintainedModel.defer_autoupdates(
+        pre_mass_update_func=disable_caching_updates,
+        post_mass_update_func=enable_caching_updates,
+    )
     def load_sample_table(self, data):
-        MaintainedModel.disable_autoupdates()
-        if self.dry_run:
-            # Don't let any auto-updates buffer because we're not going to perform the mass auto-update
-            MaintainedModel.disable_buffering()
+        # Chaching updates are not necessary when just adding data, so disabling dramatically speeds things up
         disable_caching_updates()
-        # Only auto-update fields whose update_label in the decorator is "name"
-        MaintainedModel.init_autoupdate_label_filters(label_filters=["name"])
 
         try:
             saved_aes = None
+
             with transaction.atomic():
                 try:
-                    self.load_data(data)
+                    self._load_data(data)
                 except AggregatedErrors as aes:
-                    if not self.validate and not self.defer_rollback:
+                    if self.defer_rollback:
+                        saved_aes = aes
+                    else:
                         # If we're not working for the validation interface, raise here to cause a rollback
                         raise aes
-                    else:
-                        saved_aes = aes
-            if (self.validate or self.defer_rollback) and saved_aes:
-                # If we're working for the validation interface, raise here to not cause a rollback (so that the
-                # accucor loader can be run to find more issues - samples must be loaded already to run the accucor
-                # loader), and provide the validation interface details on the exceptions.
+
+            # If we were directed to defer rollback in the event of an error, raise the exception here (outside of
+            # the atomic transaction block).  This assumes that the caller is handling rollback in their own atomic
+            # transaction blocl.
+            if saved_aes is not None:
                 raise saved_aes
 
         except Exception as e:
-            # If we're stopping with an exception, we need to clear the update buffer so that the next call doesn't
-            # make auto-updates on non-existent (or incorrect) records
-            MaintainedModel.clear_update_buffer()
-            # Re-initialize label filters to default
-            MaintainedModel.init_autoupdate_label_filters()
             enable_caching_updates()
-            MaintainedModel.enable_autoupdates()
-            if self.dry_run:
-                MaintainedModel.enable_buffering()
             raise e
 
-        # Re-initialize label filters to default
-        MaintainedModel.init_autoupdate_label_filters()
         enable_caching_updates()
-        MaintainedModel.enable_autoupdates()
-        if self.dry_run:
-            MaintainedModel.enable_buffering()
 
-    def load_data(self, data):
+    def _load_data(self, data):
         # Create a list to hold the csv reader data so that iterations from validating doesn't leave the csv reader
         # empty/at-the-end upon the import loop
         sample_table_data = list(data)
@@ -337,18 +323,6 @@ class SampleTableLoader:
             animal_rec.delete_related_caches()
         if self.verbosity >= 2:
             print("Expiring done.")
-
-        autoupdate_mode = not self.defer_autoupdates
-
-        if autoupdate_mode:
-            # No longer any need to explicitly filter based on labels, because only records containing fields with the
-            # required labels are buffered now, and when the records are buffered, the label filtering that was in
-            # effect at the time of buffering is saved so that only the fields matching the label filter will be
-            # updated.  There are autoupdates for fields in Animal and Sample, but they're only needed for FCirc
-            # calculations and will be triggered by a subsequent accucor load.
-            MaintainedModel.perform_buffered_updates()
-            # Since we only updated some of the buffered items, clear the rest of the buffer
-            MaintainedModel.clear_update_buffer()
 
     def get_tissue(self, rownum, row):
         tissue_name = self.getRowVal(row, "TISSUE_NAME")