+++ DataRepo/views/loading/validation.py
@@ -154,6 +154,8 @@ class DataValidationView(FormView):
 
         return valid, results, exceptions, ordered_keys
 
+    # No need to disable autoupdates adding the @no_autoupdates decorator to this function because supplying
+    # `validate=True` automatically disables them
     def validate_study(self):
         tmpdir_obj = tempfile.TemporaryDirectory()
         tmpdir = tmpdir_obj.name