+++ DataRepo/views/models/__init__.py
@@ -1,4 +1,5 @@
 from .animal import AnimalDetailView, AnimalListView
+from .archive_file import ArchiveFileDetailView, ArchiveFileListView
 from .compound import CompoundDetailView, CompoundListView
 from .infusate import InfusateDetailView, InfusateListView
 from .lcmethod import LCMethodDetailView, LCMethodListView
@@ -16,6 +17,8 @@ from .study import StudyDetailView, StudyListView, study_summary
 from .tissue import TissueDetailView, TissueListView
 
 __all__ = [
+    "ArchiveFileDetailView",
+    "ArchiveFileListView",
     "CompoundListView",
     "CompoundDetailView",
     "InfusateListView",